requirements.txt
fastapi[standard]
tensorflow
tensorflow-cpu
numpy
pillow
python-multipart
